Implemented QueueGame but not exactly

+++ src/ru/mirea/bert7438/pyanica/Card.java
@@ -1,9 +1,9 @@
 package ru.mirea.bert7438.pyanica;
 
 public class Card implements Comparable<Card>{
-    private final byte value;
+    private final int value;
 
-    public Card(byte value) throws Exception {
+    public Card(int value) throws Exception {
         if (value >= 0 && value<=9)
             this.value = value;
         else throw new Exception("Card must have value from 0 to 9");
@@ -17,7 +17,7 @@ public class Card implements Comparable<Card>{
     }
 
     public Card(){
-        this.value = (byte) (Math.random()*9);
+        this.value = (int) (Math.random()*9);
     }
 
     @Override
@@ -28,6 +28,6 @@ public class Card implements Comparable<Card>{
         else if (this.value == 9){
             if (o.value == 0) return -1;
         }
-        return Byte.compare(this.value, o.value);
+        return Integer.compare(this.value, o.value);
     }
 }
